pyproject.toml
[tool.poetry]
name = "music"
version = "0.1.0"
description = ""
authors = ["Stepan Kuzin <stepankuzinmail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
sentry-sdk = "^1.3.1"
requests = "^2.26.0"
alembic = "^1.7.4"
importlib-resources = "^5.4.0"
importlib-metadata = "^4.8.1"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.2"
black = "^21.7b0"
isort = "^5.9.3"
rope = "^0.19.0"
flake8-isort = "^4.0.0"
mypy = "^0.910"
sqlalchemy-stubs = "^0.4"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
